Have symptomatic individuals demand a PCR test.
