+++ tests/test_testing_models.py
@@ -4,6 +4,7 @@ import pytest
 from src.testing.testing_models import (
     _calculate_positive_tests_to_distribute_per_age_group,
 )
+from src.testing.testing_models import _request_pcr_test_bc_of_symptoms
 from src.testing.testing_models import _scale_demand_up_or_down
 from src.testing.testing_models import allocate_tests
 from src.testing.testing_models import demand_test
@@ -134,3 +135,30 @@ def test_demand_test(states, params):
     params.loc["share_known_cases"] = 1.0
 
     assert False, "`demand_test` is not tested at the moment."
+
+
+# ----------------------------------------------------------------------------
+
+
+@pytest.fixture
+def symptom_states():
+    states = pd.DataFrame()
+    # 0th and1st didn't get symptoms yesterday
+    # 2nd has a pending test, 3rd knows she's immune
+    # 4th requests a test
+    states["cd_symptoms_true"] = [-5, 3, -1, -1, -1]
+    states["pending_test"] = [False, False, True, False, False]
+    states["knows_immune"] = [False, False, False, True, False]
+    return states
+
+
+def test_request_pcr_test_bc_of_symptoms_no_one(symptom_states):
+    res = _request_pcr_test_bc_of_symptoms(symptom_states, 0.0)
+    expected = pd.Series(False, index=symptom_states.index)
+    pd.testing.assert_series_equal(res, expected)
+
+
+def test_reqst_pcr_test_bc_of_symptoms_everyone(symptom_states):
+    res = _request_pcr_test_bc_of_symptoms(symptom_states, 1.0)
+    expected = pd.Series([False, False, False, False, True], index=symptom_states.index)
+    pd.testing.assert_series_equal(res, expected)
